First implementation of movement-related hunger and API
git-svn-id: http://mc-server.googlecode.com/svn/trunk@1673 0a769ca7-a7f5-676a-18bf-c427514a06d6

+++ source/Protocol/Protocol125.cpp
@@ -469,7 +469,7 @@ void cProtocol125::SendHealth(void)
 	WriteByte (PACKET_UPDATE_HEALTH);
 	WriteShort((short)m_Client->GetPlayer()->GetHealth());
 	WriteShort(m_Client->GetPlayer()->GetFoodLevel());
-	WriteFloat(m_Client->GetPlayer()->GetFoodSaturationLevel());
+	WriteFloat((float)m_Client->GetPlayer()->GetFoodSaturationLevel());
 	Flush();
 }
